+++ CHANGES.md
@@ -1,4 +1,9 @@
 # Release History
 
+## 0.2.0
+- rename `UserInfo` to `BasicUserInfo`
+- fix use of `UnauthorizedException`
+- add example in `example` directory
+
 ## 0.1.0 
 - initial release